Move input_parser out of the lnurl module
