+++ libs/sdk-core/src/binding.rs
@@ -1,6 +1,6 @@
 use crate::breez_services::{self, BreezEvent, BreezServicesBuilder, EventListener};
 use crate::fiat::{FiatCurrency, Rate};
-use crate::lnurl::input_parser::LnUrlPayRequestData;
+use crate::input_parser::LnUrlPayRequestData;
 use crate::lsp::LspInformation;
 use crate::models::LogEntry;
 use anyhow::{anyhow, Result};
@@ -18,8 +18,8 @@ use crate::models::{
     SwapInfo,
 };
 
+use crate::input_parser::InputType;
 use crate::invoice::{self};
-use crate::lnurl::input_parser::InputType;
 use crate::lnurl::pay::model::LnUrlPayResult;
 use bip39::{Language, Mnemonic, Seed};
 
@@ -294,7 +294,7 @@ pub fn parse_invoice(invoice: String) -> Result<LNInvoice> {
 }
 
 pub fn parse(s: String) -> Result<InputType> {
-    block_on(async { crate::lnurl::input_parser::parse(&s).await })
+    block_on(async { crate::input_parser::parse(&s).await })
 }
 
 /// Second step of LNURL-pay. The first step is `parse()`, which also validates the LNURL destination
