+++ libs/sdk-core/src/breez_services.rs
@@ -5,8 +5,8 @@ use crate::grpc::channel_opener_client::ChannelOpenerClient;
 use crate::grpc::fund_manager_client::FundManagerClient;
 use crate::grpc::information_client::InformationClient;
 use crate::grpc::PaymentInformation;
+use crate::input_parser::LnUrlPayRequestData;
 use crate::invoice::{add_routing_hints, parse_invoice, LNInvoice, RouteHint, RouteHintHop};
-use crate::lnurl::input_parser::LnUrlPayRequestData;
 use crate::lnurl::pay::model::{LnUrlPayResult, ValidatedCallbackResponse};
 use crate::lnurl::pay::validate_lnurl_pay;
 use crate::lsp::LspInformation;
