+++ libs/sdk-core/src/bridge_generated.rs
@@ -27,15 +27,15 @@ use crate::fiat::LocaleOverrides;
 use crate::fiat::LocalizedName;
 use crate::fiat::Rate;
 use crate::fiat::Symbol;
+use crate::input_parser::BitcoinAddressData;
+use crate::input_parser::InputType;
+use crate::input_parser::LnUrlAuthRequestData;
+use crate::input_parser::LnUrlErrorData;
+use crate::input_parser::LnUrlPayRequestData;
+use crate::input_parser::LnUrlWithdrawRequestData;
 use crate::invoice::LNInvoice;
 use crate::invoice::RouteHint;
 use crate::invoice::RouteHintHop;
-use crate::lnurl::input_parser::BitcoinAddressData;
-use crate::lnurl::input_parser::InputType;
-use crate::lnurl::input_parser::LnUrlAuthRequestData;
-use crate::lnurl::input_parser::LnUrlErrorData;
-use crate::lnurl::input_parser::LnUrlPayRequestData;
-use crate::lnurl::input_parser::LnUrlWithdrawRequestData;
 use crate::lnurl::pay::model::LnUrlPayResult;
 use crate::lnurl::pay::model::MessageSuccessActionData;
 use crate::lnurl::pay::model::SuccessAction;
